+++ factory/def.go
@@ -3,8 +3,8 @@ package factory
 import (
 	"github.com/go-kid/ioc/component_definition"
 	"github.com/go-kid/ioc/configure"
+	"github.com/go-kid/ioc/factory/processors"
 	"github.com/go-kid/ioc/factory/support"
-	"github.com/go-kid/ioc/processors"
 )
 
 type MetaFunc func(m *component_definition.Meta) error
