+++ factory/factory.go
@@ -5,8 +5,8 @@ import (
 	"github.com/go-kid/ioc/component_definition"
 	"github.com/go-kid/ioc/configure"
 	"github.com/go-kid/ioc/definition"
+	"github.com/go-kid/ioc/factory/processors"
 	"github.com/go-kid/ioc/factory/support"
-	"github.com/go-kid/ioc/processors"
 	"github.com/go-kid/ioc/syslog"
 	"github.com/go-kid/ioc/util/reflectx"
 	"github.com/go-kid/ioc/util/sort2"
