+++ factory/processors/instantiation_aware_component_post_processors/dependency_name_aware_post_processors.go
@@ -1,6 +1,6 @@
 package instantiation_aware_component_post_processors
 
-import "github.com/go-kid/ioc/processors"
+import "github.com/go-kid/ioc/factory/processors"
 
 type dependencyNameAwarePostProcessors struct {
 	processors.DefaultInstantiationAwareComponentPostProcessor
