+++ factory/processors/instantiation_aware_component_post_processors/required_arg_validate_post_processors.go
@@ -3,7 +3,7 @@ package instantiation_aware_component_post_processors
 import (
 	"fmt"
 	"github.com/go-kid/ioc/component_definition"
-	"github.com/go-kid/ioc/processors"
+	"github.com/go-kid/ioc/factory/processors"
 )
 
 type requiredArgValidatePostProcessors struct {
