+++ factory/processors/instantiation_aware_component_post_processors/validate_aware_post_processors.go
@@ -2,7 +2,7 @@ package instantiation_aware_component_post_processors
 
 import (
 	"github.com/go-kid/ioc/component_definition"
-	"github.com/go-kid/ioc/processors"
+	"github.com/go-kid/ioc/factory/processors"
 	"github.com/go-playground/validator/v10"
 	"reflect"
 )
